Revert "Put slider example in examples"
