--- App/__init__.py
@@ -1,3 +1,0 @@
-from .buildgui import Ourcoolapp
-from .dataframe import DataFrame
-from .plotter import Plotter
