+++ general.py
@@ -1,16 +1,11 @@
+import os
+
 print('Welcome, I\'m importing a couple of modules.')
 
-from App import Ourcoolapp
-import tkinter as tk
+from dataframe import DataFrame
+from plotter import Plotter
+from buildgui import Ourcoolapp
 
-# Create tkinter window
-window = tk.Tk()
-# Call Ourcoolapp
-Ourcoolapp(window)
-# Let the window loop
-window.mainloop()
-
-'''
 # eventually this has to be some sort of open file button
 path = input('Please enter path to your file')
 
@@ -22,4 +17,4 @@ while True:
         break
     else:
         path = input(df.msg)
-'''
+
